refactor(task-holdem): reorganize code structure by moving DAO calls to a new service layer for improved maintainability
